requirements_filter.txt
numpy
pillow
tensorflow